refactor(UserDao): 리스트3-16
 - 로컬 클래스로 정의하기
